feat: implement question-manager modularization - Add 7 new custom hooks for question management - Refactor 15 components to use new hook architecture - Replace localStorage service with unified storage service - Improve code organization and maintainability - Update AI handler and realtime questions hook
