+++ src/components/ai/ai-chat-client.tsx
@@ -32,7 +32,7 @@ import VoiceAssistant from "../voice-assistant";
 
 // LocalStorage service for subjects
 class SubjectLocalStorageService {
-  private static readonly STORAGE_KEY = "exam_training_subjects";
+  private static readonly STORAGE_KEY = "akilhane_subjects";
 
   static getSubjects(): unknown[] {
     if (typeof window === "undefined") {
