+++ src/components/analytics-dashboard.tsx
@@ -138,13 +138,13 @@ export default function AnalyticsDashboard({
               : results.filter((result: QuizResult) => !result.isDemo);
 
             // Get subject information from Subjects
-            const subjects = localStorage.getItem("exam_training_subjects");
+            const subjects = localStorage.getItem("akilhane_subjects");
             const subjectsData: Subject[] = subjects
               ? JSON.parse(subjects)
               : [];
 
             // Get question information from Questions (not currently used)
-            // const questions = localStorage.getItem('exam_training_questions');
+            // const questions = localStorage.getItem('akilhane_questions');
 
             if (filteredResults.length === 0) {
               // If there are no quiz results, use simple mock data
