+++ src/components/flashcard.tsx
@@ -98,7 +98,7 @@ const FlashcardComponent: React.FC<FlashcardProps> = ({
             return [];
           }
           try {
-            const stored = localStorage.getItem("exam_training_questions");
+            const stored = localStorage.getItem("akilhane_questions");
             const questions = stored ? JSON.parse(stored) : [];
             return questions.filter(
               (q: { subject: string }) => q.subject === subject,
