+++ src/lib/subject-ai-handler.ts
@@ -28,7 +28,7 @@ interface LocalStorageSubject {
 
 // LocalStorage service for subjects
 class SubjectLocalStorageService {
-  private static readonly STORAGE_KEY = "exam_training_subjects";
+  private static readonly STORAGE_KEY = "akilhane_subjects";
 
   static getSubjects(): LocalStorageSubject[] {
     if (typeof window === "undefined") {
